workspace/python/agents/short-movie-agents/pyproject.toml
[project]
name = "short-movie-agents"
version = "0.3.1"
description = "Demo showcasing how agents can collaborate to build short movies based on user requests."
authors = [{ name = "Lastingyeh" }]
license = { text = "Apache License 2.0" }
readme = "README.md"
requires-python = ">=3.13,<3.14"

dependencies = [
    "google-adk>=1.14.0,<2.0.0",
    "opentelemetry-exporter-gcp-trace>=1.9.0",
    "google-cloud-logging>=3.12.0",
    "google-cloud-aiplatform[evaluation]>=1.116.0",
    "fastapi>=0.115.8",
    "uvicorn>=0.34.0",
    "psycopg2-binary>=2.9.10",
]

[dependency-groups]
dev = [
    "pytest>=8.3.4",
    "pytest-asyncio>=0.23.8",
    "nest-asyncio>=1.6.0",
    "mypy>=1.14.0",
    "jupyter>=1.0.0",
]

[tool.ruff]
# Point back to the Master template in the root
extend = "../../../pyproject.toml"

[tool.ruff.lint.isort]
known-first-party = ["app"]

[tool.mypy]
disallow_untyped_calls = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
no_implicit_optional = true
check_untyped_defs = true
disallow_subclassing_any = true
warn_incomplete_stub = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_unreachable = true
follow_imports = "silent"
ignore_missing_imports = true
explicit_package_bases = true
disable_error_code = ["misc", "no-untyped-call", "no-any-return"]
exclude = [".venv"]

[tool.pytest.ini_options]
pythonpath = "."
asyncio_default_fixture_loop_scope = "function"

[tool.hatch.build.targets.wheel]
packages = ["app"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
